Add test for rrule with different display timezone
The test creates an `RRule` instance with a start date that has
the timezone **Europe/Berlin**. ITs sserializes and encodes it to
a json string. Then it creates a new `RRule` object from this
json string. The new `RRule` gets **America/New_York** as display
timezone and `$firstHandoff` as `DateTime` object as start date
that is the same time as the start date in the json but in another
timezone. Then the tests creates 7 recurrences and asserts for each
that the timezone is the display timezone and that the dates are
the expected dates.